Refactor namespaces and class structures in several domain files
The commit includes refactoring the structure of namespaces and classes in multiple domain files, such as ActionLog, ActionType, EntityBase, Position, Rack, Material, SizeType, and Stock. This is to improve readability and coherence with a simplified structure using C# 10 file-scoped namespaces. A couple of unnecessary comments and using statements have also been removed.