Add ResultPattern project to solution
A new project called ResultPattern has been added to the StockFlow solution. This encapsulates the result of an operation which could either be a success or failure. Accordingly, new classes and interfaces have been added in this project. References to this project have also been added in the Infrastructure and Application project files.