Update project file and application services references
Project references and package versions have been updated in the solution files. Changes also include the addition of the Application service in Program.cs and adjustments to the rendermode in Counter.razor. Details of references and versions are now more concise and finely arranged. All modifications aim to improve code readability and maintainability.